requirements.txt
streamlit
ultralytics
pillow
pandas
requests
